requirement.txt
requests
mysql-connector-python
mariadb
mysqlclient
